feat(core): refactor codebase to modular structure with full feature access
- Introduced modular app structure with packages: core, models, services, utils, auth, api, main
- Added centralized configuration management in core/config.py
- Implemented application factory in core/factory.py for flexible app creation
- Created detailed models for User, Tutorial, UsageLog, and APIKey with full attributes
- Developed AI tutorial generation service with free access to all features in services/ai.py
- Integrated OpenRouter API for text and audio generation in services/openrouter.py
- Added Kyutai TTS streaming service in services/tts.py for real-time audio
- Enhanced helpers with usage tracking, rate limiting, validation, and analytics in utils/helpers.py
- Updated main app entry point to app_main.py and adjusted Procfile accordingly
- Added comprehensive test runner script run_tests.py
- Included production WSGI entry point wsgi.py
- Removed redundant files and large temporary assets, saving ~35MB
- Improved code quality, maintainability, testability, and scalability
- Enabled all users to have premium content and unlimited access

This major refactor prepares the codebase for scalable development and deployment with enhanced AI-powered tutorial generation and audio support.
