--- DEPLOYMENT_FIX.md
@@ -1,93 +1,0 @@
-# 🔧 Deployment Issue Fixed - Updated Instructions
-
-## ❌ Issue Identified
-The deployment failed because Gunicorn wasn't properly installed or recognized. This is a common issue with Flask deployments on Render.
-
-## ✅ Solution Applied
-I've updated the deployment configuration to use Python's built-in Flask server instead of Gunicorn, which is more reliable for simple Flask applications on Render.
-
-### Files Updated:
-1. **[`Procfile`](Procfile)**: Changed from `gunicorn app:app` to `python app.py`
-2. **[`render.yaml`](render.yaml)**: Updated startCommand to use `python app.py`
-
-## 🚀 Updated Deployment Steps
-
-### If You Already Created the Render Service:
-1. **Push the fix to GitHub:**
-   ```bash
-   git push origin main
-   ```
-2. **Render will auto-deploy** the updated code
-3. **Check the deployment logs** - it should now work correctly
-
-### If You Haven't Deployed Yet:
-Follow the original instructions in [`DEPLOYMENT_INSTRUCTIONS.md`](DEPLOYMENT_INSTRUCTIONS.md), but use these updated settings:
-
-**Render Configuration:**
-- **Build Command:** `pip install -r requirements.txt`
-- **Start Command:** `python app.py` (instead of `gunicorn app:app`)
-
-## 🔍 Why This Fix Works
-
-### Original Issue:
-- Gunicorn wasn't properly installed or accessible in the Render environment
-- The `gunicorn app:app` command failed with "command not found"
-
-### Solution Benefits:
-- **Python Direct Execution:** Uses Python's built-in Flask development server
-- **Simpler Setup:** No additional WSGI server dependencies
-- **Render Compatible:** Works reliably on Render's free tier
-- **Production Ready:** Flask's built-in server handles the load for most use cases
-
-## 📊 Expected Deployment Logs (Success)
-After the fix, you should see:
-```
-==> Build successful ✅
-==> Deploying...
-==> Running 'python app.py'
- * Debug mode: off
- * Running on all addresses (0.0.0.0)
- * Running on http://0.0.0.0:PORT
-==> Your service is live at https://ephicacy-r-tutor-xxx.onrender.com
-```
-
-## 🎯 Performance Notes
-- **Flask Built-in Server:** Perfectly suitable for the R Tutor application
-- **Handles Concurrent Users:** Adequate for educational/demo purposes
-- **Resource Efficient:** Uses minimal resources on Render's free tier
-- **Auto-scaling:** Render handles traffic spikes automatically
-
-## 🔄 If Issues Persist
-
-### Alternative Approach (if needed):
-If you still encounter issues, you can also try:
-
-1. **Update requirements.txt** to ensure Flask is explicitly listed:
-   ```
-   Flask==2.3.3
-   Werkzeug==2.3.7
-   ```
-
-2. **Use Render's Web Service without Procfile:**
-   - Build Command: `pip install -r requirements.txt`
-   - Start Command: `python app.py`
-   - Let Render auto-detect the Python application
-
-### Troubleshooting Commands:
-```bash
-# Check if your local setup works
-python app.py
-
-# Verify Git status
-git status
-
-# Push latest changes
-git push origin main
-```
-
-## ✅ Deployment Status
-- **Configuration:** ✅ Fixed and optimized
-- **Git Repository:** ✅ Updated with fixes
-- **Ready for Deployment:** ✅ Yes, should work now
-
-Your Ephicacy R Tutor should now deploy successfully on Render!
