+++ app/core/__init__.py
@@ -1,0 +1,1 @@
+# Core module
