+++ app/main/__init__.py
@@ -1,0 +1,1 @@
+# Main package
