+++ app/models/__init__.py
@@ -1,0 +1,1 @@
+# Models package\ntry:\n    from ..core.extensions import db\n    from .user import User\n    from .tutorial import Tutorial\n    from .usage import UsageLog, APIKey\nexcept ImportError:\n    # Handle cases where Flask dependencies aren't available\n    pass\n\n# Database initialization function\ndef init_db(app):\n    \"\"\"Initialize database with app context\"\"\"\n    db.init_app(app)\n    \n    with app.app_context():\n        db.create_all()\n        \n        # Create indexes for better performance\n        try:\n            with db.engine.connect() as conn:\n                conn.execute(\n                    db.text('CREATE INDEX IF NOT EXISTS idx_usage_user_action_date ON usage_logs(user_id, action, created_at)')\n                )\n                conn.execute(\n                    db.text('CREATE INDEX IF NOT EXISTS idx_tutorials_user_date ON tutorials(user_id, created_at)')\n                )\n            print(\"Database indexes created successfully\")\n        except Exception as e:\n            print(f\"Index creation skipped: {e}\")\n        \n        print(\"Database initialized successfully - OpenRouter integration enabled\")\n\n__all__ = ['db', 'User', 'Tutorial', 'UsageLog', 'APIKey', 'init_db']
