--- deployment_config.txt
@@ -1,299 +1,0 @@
-# requirements.txt
-flask==2.3.3
-flask-sqlalchemy==3.0.5
-flask-login==0.6.3
-werkzeug==2.3.7
-requests==2.31.0
-python-dotenv==1.0.0
-gunicorn==21.2.0
-psycopg2-binary==2.9.7
-redis==4.6.0
-pyttsx3==2.90
-pydub==0.25.1
-
----
-# Dockerfile
-FROM python:3.11-slim
-
-WORKDIR /app
-
-# Install system dependencies
-RUN apt-get update && apt-get install -y \
-    gcc \
-    g++ \
-    libasound2-dev \
-    libportaudio2 \
-    libportaudiocpp0 \
-    portaudio19-dev \
-    && rm -rf /var/lib/apt/lists/*
-
-# Copy requirements and install Python deps
-COPY requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
-
-# Copy application code
-COPY . .
-
-# Create directories for uploads and logs
-RUN mkdir -p static/generated_audio logs
-
-# Expose port
-EXPOSE 5000
-
-# Health check
-HEALTHCHECK --interval=30s --timeout=10s --start-period=5s --retries=3 \
-    CMD curl -f http://localhost:5000/health || exit 1
-
-# Run with gunicorn for production
-CMD ["gunicorn", "--bind", "0.0.0.0:5000", "--workers", "2", "--timeout", "120", "app:app"]
-
----
-# docker-compose.yml
-version: '3.8'
-
-services:
-  web:
-    build: .
-    ports:
-      - "5000:5000"
-    environment:
-      - DATABASE_URL=postgresql://rtutoruser:rtutorpass@db:5432/rtutor_saas
-      - REDIS_URL=redis://redis:6379
-      - SECRET_KEY=${SECRET_KEY:-dev-secret-change-in-production}
-      - DEEPSEEK_API_KEY=${DEEPSEEK_API_KEY}
-      - OPENAI_API_KEY=${OPENAI_API_KEY}
-    depends_on:
-      - db
-      - redis
-    volumes:
-      - ./static/generated_audio:/app/static/generated_audio
-      - ./logs:/app/logs
-    restart: unless-stopped
-
-  db:
-    image: postgres:15-alpine
-    environment:
-      - POSTGRES_DB=rtutor_saas
-      - POSTGRES_USER=rtutoruser
-      - POSTGRES_PASSWORD=rtutorpass
-    volumes:
-      - postgres_data:/var/lib/postgresql/data
-    restart: unless-stopped
-
-  redis:
-    image: redis:7-alpine
-    volumes:
-      - redis_data:/data
-    restart: unless-stopped
-
-  nginx:
-    image: nginx:alpine
-    ports:
-      - "80:80"
-      - "443:443"
-    volumes:
-      - ./nginx.conf:/etc/nginx/nginx.conf
-      - ./ssl:/etc/nginx/ssl
-    depends_on:
-      - web
-    restart: unless-stopped
-
-volumes:
-  postgres_data:
-  redis_data:
-
----
-# nginx.conf
-events {
-    worker_connections 1024;
-}
-
-http {
-    upstream app {
-        server web:5000;
-    }
-
-    # Rate limiting
-    limit_req_zone $binary_remote_addr zone=api:10m rate=10r/m;
-    limit_req_zone $binary_remote_addr zone=auth:10m rate=5r/m;
-
-    server {
-        listen 80;
-        server_name localhost;
-
-        # Security headers
-        add_header X-Frame-Options DENY;
-        add_header X-Content-Type-Options nosniff;
-        add_header X-XSS-Protection "1; mode=block";
-
-        # Static files
-        location /static/ {
-            alias /app/static/;
-            expires 1y;
-            add_header Cache-Control "public, immutable";
-        }
-
-        # API rate limiting
-        location /generate-tutorial {
-            limit_req zone=api burst=5 nodelay;
-            proxy_pass http://app;
-            proxy_set_header Host $host;
-            proxy_set_header X-Real-IP $remote_addr;
-            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-            proxy_set_header X-Forwarded-Proto $scheme;
-        }
-
-        # Auth rate limiting
-        location ~ ^/(login|register) {
-            limit_req zone=auth burst=3 nodelay;
-            proxy_pass http://app;
-            proxy_set_header Host $host;
-            proxy_set_header X-Real-IP $remote_addr;
-            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-            proxy_set_header X-Forwarded-Proto $scheme;
-        }
-
-        # All other requests
-        location / {
-            proxy_pass http://app;
-            proxy_set_header Host $host;
-            proxy_set_header X-Real-IP $remote_addr;
-            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-            proxy_set_header X-Forwarded-Proto $scheme;
-        }
-    }
-}
-
----
-# .env.example
-# Database Configuration
-DATABASE_URL=postgresql://username:password@localhost:5432/rtutor_saas
-
-# Redis Configuration (for caching and sessions)
-REDIS_URL=redis://localhost:6379
-
-# Application Security
-SECRET_KEY=your-super-secret-key-here
-
-# AI API Keys (Optional - for premium features)
-DEEPSEEK_API_KEY=your-deepseek-api-key
-OPENAI_API_KEY=your-openai-api-key
-
-# Email Configuration (for notifications)
-SMTP_SERVER=smtp.gmail.com
-SMTP_PORT=587
-SMTP_USERNAME=your-email@gmail.com
-SMTP_PASSWORD=your-app-password
-
-# Stripe Configuration (for payments)
-STRIPE_PUBLISHABLE_KEY=pk_test_...
-STRIPE_SECRET_KEY=sk_test_...
-STRIPE_WEBHOOK_SECRET=whsec_...
-
-# Environment
-ENVIRONMENT=production
-DEBUG=False
-
----
-# GitHub Actions Workflow (.github/workflows/deploy.yml)
-name: Deploy R Tutor SaaS
-
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    
-    services:
-      postgres:
-        image: postgres:15
-        env:
-          POSTGRES_PASSWORD: postgres
-          POSTGRES_DB: test_db
-        options: >-
-          --health-cmd pg_isready
-          --health-interval 10s
-          --health-timeout 5s
-          --health-retries 5
-        ports:
-          - 5432:5432
-
-    steps:
-    - uses: actions/checkout@v3
-    
-    - name: Set up Python
-      uses: actions/setup-python@v4
-      with:
-        python-version: '3.11'
-    
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install -r requirements.txt
-        pip install pytest pytest-flask
-    
-    - name: Run tests
-      env:
-        DATABASE_URL: postgresql://postgres:postgres@localhost:5432/test_db
-        SECRET_KEY: test-secret-key
-      run: |
-        pytest tests/ -v
-
-  deploy:
-    needs: test
-    runs-on: ubuntu-latest
-    if: github.ref == 'refs/heads/main'
-    
-    steps:
-    - uses: actions/checkout@v3
-    
-    - name: Deploy to production
-      run: |
-        echo "Add your deployment script here"
-        # Example: Deploy to AWS ECS, Google Cloud Run, or DigitalOcean
-
----
-# Local development script (run.py)
-#!/usr/bin/env python3
-"""
-Local development server with hot reload and debugging
-"""
-import os
-import sys
-from dotenv import load_dotenv
-
-# Load environment variables
-load_dotenv()
-
-# Add current directory to path
-sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
-
-from app import app, db
-
-def init_db():
-    """Initialize database with sample data"""
-    with app.app_context():
-        db.create_all()
-        print("✅ Database initialized")
-
-if __name__ == '__main__':
-    print("🚀 Starting R Tutor SaaS - Development Mode")
-    print("📊 Dashboard: http://localhost:5000")
-    print("🔧 Set environment variables in .env file")
-    
-    # Initialize database on first run
-    if '--init-db' in sys.argv:
-        init_db()
-        sys.exit(0)
-    
-    # Start development server
-    app.run(
-        debug=True,
-        host='0.0.0.0',
-        port=int(os.environ.get('PORT', 5000)),
-        use_reloader=True
-    )
